backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
openai==1.93.0
anthropic>=0.49.0
google-genai>=1.14.0
pydantic==2.10.2
pydantic-settings==2.6.1
cachetools==4.2.2
python-multipart==0.0.12
pandas==2.2.3
python-dotenv==1.0.1
slowapi==0.1.9
python-json-logger==3.2.1
sentry-sdk[fastapi]==2.19.2
tenacity==9.1.2
pytest==8.4.1
